kube: move clusterStore to its own file
